requirements.txt
flask
flask-login
flask-migrate
flask-sqlalchemy
Flask-Migrate